+++ CHANGELOG.md
@@ -8,6 +8,8 @@ structure fields anymore. So that your data structures don't end up with
 errors in it, without knowing where they were actually assigned.
 * Bugfix: Assigning to lists and maps inside a std:to\_drop marked value
 works now too.
+* Doc: Documented closures briefly, reference data types and how to do
+object oriented programming.
 
 0.3.4 (2019-11-20)
 ==================